+++ CHANGELOG.md
@@ -71,6 +71,8 @@
 
 ### 🚜 Refactor
 
+- *(common,lib)* Remove or standardise some more code - ([76761f5](https://github.com/JayRovacsek/nix-config/commit/76761f5fa162afb7e7dcae33ab37c9aa1424d794))
+
 - *(flake)* Remove dead config files - ([6392de3](https://github.com/JayRovacsek/nix-config/commit/6392de32ed3764010579c58bda08d0befaaf23b3))
 
 - *(flake)* Rename common.networking to common.config - ([6392de3](https://github.com/JayRovacsek/nix-config/commit/6392de32ed3764010579c58bda08d0befaaf23b3))